Develop

+++ modules/users/admin/del.php
@@ -26,7 +26,7 @@ if( $numrows )
 	die( "NO" );
 }
 
-$sql = "SELECT `username`, `full_name`, `email`, `photo`, `in_groups` FROM `" . $db_config['dbsystem'] . "`.`" . NV_USERS_GLOBALTABLE . "` WHERE `userid`=" . $userid;
+$sql = "SELECT `username`, `full_name`, `email`, `photo`, `idsite` FROM `" . $db_config['dbsystem'] . "`.`" . NV_USERS_GLOBALTABLE . "` WHERE `userid`=" . $userid;
 $query = $db->sql_query( $sql );
 $numrows = $db->sql_numrows( $query );
 if( $numrows != 1 )
@@ -34,43 +34,47 @@ if( $numrows != 1 )
 	die( "NO" );
 }
 
-list( $username, $full_name, $email, $photo, $in_groups ) = $db->sql_fetchrow( $query );
+list( $username, $full_name, $email, $photo, $idsite ) = $db->sql_fetchrow( $query );
 
-$userdelete = ( ! empty( $full_name ) ) ? $full_name . " (" . $username . ")" : $username;
-
-$result = $db->sql_query( "DELETE FROM `" . $db_config['dbsystem'] . "`.`" . NV_USERS_GLOBALTABLE . "` WHERE `userid`=" . $userid );
-if( ! $result )
+if( $global_config['idsite'] > 0 AND $idsite != $global_config['idsite'] )
 {
 	die( "NO" );
 }
 
-$result = $db->sql_query( "DELETE FROM `" . $db_config['dbsystem'] . "`.`" . NV_USERS_GLOBALTABLE . "_openid` WHERE `userid`=" . $userid );
-if( ! empty( $in_groups ) )
+$query = $db->sql_query( "SELECT * FROM `" . $db_config['dbsystem'] . "`.`" . NV_GROUPS_GLOBALTABLE . "_users` WHERE `group_id` IN (1,2,3) AND `userid`=" . $userid );
+if( $db->sql_numrows( $query ) )
 {
-	$result = $db->sql_query( "SELECT `group_id`, `users` FROM `" . NV_GROUPS_GLOBALTABLE . "` WHERE `group_id` IN (" . $in_groups . ")" );
-	while( list( $group_id, $users ) = $db->sql_fetchrow( $result ) )
+	die( "ERROR_" . $lang_module['delete_group_system'] );
+}
+else
+{
+	$userdelete = ( ! empty( $full_name ) ) ? $full_name . " (" . $username . ")" : $username;
+
+	$result = $db->sql_query( "DELETE FROM `" . $db_config['dbsystem'] . "`.`" . NV_USERS_GLOBALTABLE . "` WHERE `userid`=" . $userid );
+	if( ! $result )
 	{
-		$users = "," . $users . ",";
-		$users = str_replace( "," . $userid . ",", ",", $users );
-		$users = trim( $users, "," );
-		$db->sql_query( "UPDATE `" . NV_GROUPS_GLOBALTABLE . "` SET `users` = '" . $users . "' WHERE `group_id`=" . $group_id );
+		die( "NO" );
 	}
+
+	$db->sql_query( "UPDATE `" . $db_config['dbsystem'] . "`.`" . NV_GROUPS_GLOBALTABLE . "` SET `number` = `number`-1 WHERE `group_id` IN (SELECT `group_id` FROM `" . $db_config['dbsystem'] . "`.`" . NV_GROUPS_GLOBALTABLE . "_users` WHERE `userid`=" . $userid . ")" );
+	$db->sql_query( "DELETE FROM `" . $db_config['dbsystem'] . "`.`" . NV_GROUPS_GLOBALTABLE . "_users` WHERE `userid`=" . $userid );
+	$db->sql_query( "DELETE FROM `" . $db_config['dbsystem'] . "`.`" . NV_USERS_GLOBALTABLE . "_openid` WHERE `userid`=" . $userid );
+	$db->sql_query( "DELETE FROM `" . $db_config['dbsystem'] . "`.`" . NV_USERS_GLOBALTABLE . "_info` WHERE `userid`=" . $userid );
+
+	nv_insert_logs( NV_LANG_DATA, $module_name, 'log_del_user', "userid " . $userid, $admin_info['userid'] );
+
+	if( ! empty( $photo ) and is_file( NV_ROOTDIR . '/' . $photo ) )
+	{
+		@nv_deletefile( NV_ROOTDIR . '/' . $photo );
+	}
+
+	$subject = $lang_module['delconfirm_email_title'];
+	$message = sprintf( $lang_module['delconfirm_email_content'], $userdelete, $global_config['site_name'] );
+	$message = str_replace( "\n", "<br />", $message );
+	$message .= "<br /><br />------------------------------------------------<br /><br />";
+	$message .= nv_EncString( $message );
+	nv_sendmail( $global_config['site_email'], $email, $subject, $message );
+	die( "OK" );
 }
-
-nv_insert_logs( NV_LANG_DATA, $module_name, 'log_del_user', "userid " . $userid, $admin_info['userid'] );
-
-if( ! empty( $photo ) and is_file( NV_ROOTDIR . '/' . $photo ) )
-{
-	@nv_deletefile( NV_ROOTDIR . '/' . $photo );
-}
-
-$subject = $lang_module['delconfirm_email_title'];
-$message = sprintf( $lang_module['delconfirm_email_content'], $userdelete, $global_config['site_name'] );
-$message = str_replace( "\n", "<br />", $message );
-$message .= "<br /><br />------------------------------------------------<br /><br />";
-$message .= nv_EncString( $message );
-nv_sendmail( $global_config['site_email'], $email, $subject, $message );
-
-die( "OK" );
 
 ?>
